requirements.txt
python-binance==1.0.19
streamlit==1.39.0